requirements.txt
Flask==1.1.2
Flask_SocketIO
Flask-User==0.6.21
itsdangerous==1.1.0
Jinja2==2.11.3
MarkupSafe==1.1.1
python-engineio
python-socketio
six==1.15.0
Werkzeug==1.0.1
Flask-login==0.5.0
gevent-websocket
Flask-Security==3.0.0
pystache=0.5.4